requirements.txt
streamlit>=1.28.0
openai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
st_img_pastebutton>=0.0.5
boto3>=1.28.0